fix(material/tree): Start aria-level numbering at 1, not 0
I started by trying to change the value of the `level` variable to start
at 1 [0], but in my application, that prevented all nodes from being
rendered. I didn't look too closely at that, but it seems like it
might be convenient for the code representation of the node to start at 0.

So becuase of that, this commit changes the aria-level value only.

[0]:
https://github.com/angular/components/blob/db879d5/src/material/tree/data-source/flat-data-source.ts#L92